Extract in-game map editor object handling into Editor
Move EditObjects (the editor's place/edit-world-objects routine, ~260 lines)
out of ZzzInterface.cpp into Engine/Object/EditObjects.{h,cpp} under namespace
Editor. Its single caller (MainScene) is now qualified; the editor brush/
selection globals it reads stay in ZzzInterface.cpp via extern. No behaviour
change.
