+++ src/source/Engine/Object/EditObjects.h
@@ -1,0 +1,8 @@
+#pragma once
+
+// In-game map editor: places/edits world objects, monsters and terrain mapping
+// from the editor's current brush/selection. Extracted from ZzzInterface.cpp.
+namespace Editor
+{
+    void EditObjects();
+}
