+++ src/source/Engine/Object/ZzzInterface.h
@@ -60,7 +60,6 @@ void SendRequestAction(OBJECT& obj, BYTE action);
 
 bool PressKey(int Key);
 void MoveHero();
-void EditObjects();
 
 void ClearInput(BOOL bClearWhisperTarget = TRUE);
 void CutText(const wchar_t* Text, wchar_t* Text1, wchar_t* Text2, size_t MaxLength);
